refactor: move chatEmojis to src/util/emojis
