+++ src/events/mineflayer/handler/kicked.ts
@@ -1,4 +1,4 @@
-import Emojis from "../../../util/emojis";
+import Emojis from "../../../util/emojis/chatEmojis";
 import { Event } from "../../../interfaces/Event";
 
 export default {
